requirements.txt
fastapi 
uvicorn
sqlalchemy 
psycopg2-binary
cryptography
python-keycloak 
python-jose[cryptography]